리뷰 반영: 관리자 CSRF 템플릿과 테스트 정리
