+++ src/main/resources/templates/admin/login.html
@@ -96,7 +96,6 @@
         <div th:if="${error}" class="error" th:text="${error}"></div>
 
         <form th:action="@{/admin/login}" method="post">
-            <input type="hidden" th:name="${_csrf.parameterName}" th:value="${_csrf.token}">
             <div class="form-group">
                 <label for="username">관리자 아이디</label>
                 <input type="text" id="username" name="username"
