+++ src/main/resources/templates/admin/upload.html
@@ -120,7 +120,6 @@
             <div class="nav-links">
                 <a href="/admin/validate">검증 페이지</a>
                 <form th:action="@{/admin/logout}" method="post">
-                    <input type="hidden" th:name="${_csrf.parameterName}" th:value="${_csrf.token}">
                     <button type="submit">로그아웃</button>
                 </form>
             </div>
@@ -133,7 +132,6 @@
             <h2>📄 PDF 업로드 (수강편람)</h2>
             <p class="description">수강편람 PDF 파일을 업로드하면 Gemini AI로 파싱하여 DB에 저장합니다.</p>
             <form th:action="@{/admin/upload/pdf}" method="post" enctype="multipart/form-data">
-                <input type="hidden" th:name="${_csrf.parameterName}" th:value="${_csrf.token}">
                 <div class="file-input-wrapper">
                     <input type="file" name="file" accept=".pdf" required>
                 </div>
@@ -145,7 +143,6 @@
             <h2>📊 Excel 업로드 (종합강의시간표)</h2>
             <p class="description">종합강의시간표 Excel 파일(.xlsx)을 학수번호 기준으로 비교하여 DB에 반영합니다. 삭제 대상은 물리 삭제하지 않고 비활성화합니다.</p>
             <form th:action="@{/admin/upload/excel}" method="post" enctype="multipart/form-data">
-                <input type="hidden" th:name="${_csrf.parameterName}" th:value="${_csrf.token}">
                 <div style="margin-bottom: 15px;">
                     <label for="semester" style="display:block; margin-bottom:8px; font-weight:600;">학기</label>
                     <input id="semester" type="text" name="semester" placeholder="예: 2026-2" required
