+++ src/main/resources/templates/admin/validate-result.html
@@ -186,7 +186,6 @@
                 <a href="/admin/validate">다시 검증하기</a>
                 <a href="/admin/upload">업로드 페이지</a>
                 <form th:action="@{/admin/logout}" method="post">
-                    <input type="hidden" th:name="${_csrf.parameterName}" th:value="${_csrf.token}">
                     <button type="submit">로그아웃</button>
                 </form>
             </div>
